Recommendation system using MovieLens data
This is complete documentation for building recommendation system on the cloud, this project was done for college mini-project as a team in 2019 and it will guide you to build a recommendation system. Here, our model was able to predict movies based on individual user interest in a simple execution part.